pyproject.toml
[tool.poetry]
name = "agent-web-search"
version = "0.1.0"
description = ""
authors = ["Matt Norris <matt@mattnorris.dev>"]
readme = "README.md"
packages = [{ include = "agent_web_search", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.1.0"
python-decouple = "^3.8"
langchain-community = "^0.3.22"
langgraph = "^0.3.34"
langgraph-cli = "^0.2.7"
langchain-openai = "^0.3.14"
langchain-core = "^0.3.56"
langsmith = "^0.3.38"
pydantic = "^2.11.4"


[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "^25.1.0"}
pre-commit = "^4.2.0"
pylint = "^3.3.6"
perflint = "^0.8.1"
ruff = "^0.11.7"
bandit = "^1.8.3"
pytest = "^8.3.5"
mypy = "^1.15.0"
langgraph-cli = { extras = ["inmem"], version = "^0.2.7" }
jupyter = "^1.1.1"
detect-secrets = "^1.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
